pyproject.toml
[project]
name = "powerpoint-mcp"
version = "0.1.0"
description = "A PowerPoint MCP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp==2.12.4",
    "google-adk>=1.9.0",
    "python-pptx>=1.0.2",
    "starlette>=0.47.2",
    "pydantic>=2.0",
]
